Organization and grouping of the project folder structure.
