--- src/components/Logo.tsx
@@ -1,26 +1,0 @@
-const Logo = () => {
-  return (
-    <a className="logo" href="#">
-      <svg
-        className="icon-logo "
-        width="27"
-        height="32"
-        viewBox="0 0 27 32"
-        xmlns="http://www.w3.org/2000/svg"
-      >
-        <g clipPath="url(#clip0_372_134)">
-          <path
-            d="M10.69 9.61L18.01 28.48C18.31 29.24 18.65 29.75 19.03 30.01C19.41 30.25 19.76 30.38 20.08 30.4V31C19.68 30.96 19.19 30.94 18.61 30.94C18.03 30.92 17.45 30.91 16.87 30.91C16.09 30.91 15.36 30.92 14.68 30.94C14 30.94 13.45 30.96 13.03 31V30.4C14.05 30.36 14.7 30.19 14.98 29.89C15.26 29.57 15.2 28.9 14.8 27.88L9.22 12.97L9.7 12.58L4.48 26.14C4.16 26.94 3.97 27.62 3.91 28.18C3.85 28.72 3.9 29.15 4.06 29.47C4.24 29.79 4.53 30.02 4.93 30.16C5.35 30.3 5.87 30.38 6.49 30.4V31C5.93 30.96 5.34 30.94 4.72 30.94C4.12 30.92 3.56 30.91 3.04 30.91C2.54 30.91 2.11 30.92 1.75 30.94C1.41 30.94 1.09 30.96 0.79 31V30.4C1.19 30.3 1.6 30.07 2.02 29.71C2.44 29.33 2.82 28.69 3.16 27.79L10.21 9.61C10.29 9.61 10.37 9.61 10.45 9.61C10.53 9.61 10.61 9.61 10.69 9.61ZM14.35 22.36V22.96H5.41L5.71 22.36H14.35Z"
-            fill=""
-          />
-          <path
-            d="M18.68 0.429999C19.68 0.429999 20.44 0.549999 20.96 0.789999C21.48 1.01 21.94 1.26 22.34 1.54C22.58 1.68 22.77 1.79 22.91 1.87C23.07 1.93 23.22 1.96 23.36 1.96C23.56 1.96 23.7 1.85 23.78 1.63C23.88 1.41 23.96 1.08 24.02 0.64H24.71C24.69 0.979999 24.66 1.39 24.62 1.87C24.6 2.33 24.58 2.95 24.56 3.73C24.56 4.49 24.56 5.5 24.56 6.76H23.87C23.81 5.8 23.6 4.88 23.24 4C22.88 3.12 22.35 2.4 21.65 1.84C20.97 1.28 20.07 0.999999 18.95 0.999999C17.89 0.999999 17.01 1.32 16.31 1.96C15.63 2.6 15.29 3.44 15.29 4.48C15.29 5.38 15.52 6.14 15.98 6.76C16.44 7.36 17.05 7.91 17.81 8.41C18.59 8.89 19.43 9.41 20.33 9.97C21.37 10.61 22.29 11.25 23.09 11.89C23.91 12.51 24.55 13.2 25.01 13.96C25.49 14.72 25.73 15.64 25.73 16.72C25.73 18 25.44 19.06 24.86 19.9C24.28 20.74 23.51 21.37 22.55 21.79C21.59 22.21 20.53 22.42 19.37 22.42C18.31 22.42 17.47 22.3 16.85 22.06C16.23 21.82 15.7 21.57 15.26 21.31C14.82 21.03 14.48 20.89 14.24 20.89C14.04 20.89 13.89 21 13.79 21.22C13.71 21.44 13.64 21.77 13.58 22.21H12.89C12.93 21.79 12.95 21.3 12.95 20.74C12.97 20.16 12.98 19.41 12.98 18.49C13 17.57 13.01 16.41 13.01 15.01H13.7C13.78 16.21 14 17.33 14.36 18.37C14.74 19.41 15.31 20.25 16.07 20.89C16.85 21.51 17.88 21.82 19.16 21.82C19.82 21.82 20.44 21.69 21.02 21.43C21.62 21.15 22.11 20.72 22.49 20.14C22.87 19.54 23.06 18.77 23.06 17.83C23.06 17.01 22.86 16.3 22.46 15.7C22.08 15.08 21.54 14.51 20.84 13.99C20.14 13.45 19.32 12.9 18.38 12.34C17.44 11.76 16.56 11.17 15.74 10.57C14.92 9.97 14.26 9.28 13.76 8.5C13.28 7.7 13.04 6.74 13.04 5.62C13.04 4.44 13.3 3.47 13.82 2.71C14.36 1.93 15.06 1.36 15.92 0.999999C16.78 0.619999 17.7 0.429999 18.68 0.429999Z"
-            fill=""
-          />
-        </g>
-      </svg>
-    </a>
-  );
-};
-
-export default Logo;
