+++ src/pages/HomePage.tsx
@@ -1,0 +1,16 @@
+import Contacts from "@components/Contacts";
+import Hero from "@components/Hero";
+import Projects from "@components/Projects/Projects";
+import { Fragment } from "react";
+
+const HomePage = () => {
+  return (
+    <Fragment>
+      <Hero />
+      <Projects />
+      <Contacts />
+    </Fragment>
+  );
+};
+
+export default HomePage;
